+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-13T06:47:46.437Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-13T09:41:43.187Z")
     r"""The date and time the resource was last updated"""