+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K2H2YVRJQT1G9YV0B8FX3VB2"
+        "01K2HCXCF70FJXXJE78H0G4DKP"
     )
     r"""The id of the resource"""