+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -179,7 +179,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-13T06:47:46.437Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-13T09:41:43.187Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")